Move exceptions to dtbase/backend/exc.py
