+++ dtbase/backend/exc.py
@@ -1,0 +1,14 @@
+class RowExistsError(Exception):
+    pass
+
+
+class RowMissingError(Exception):
+    pass
+
+
+class TooManyRowsError(Exception):
+    pass
+
+
+class DatabaseConnectionError(Exception):
+    pass
