+++ dtbase/backend/routers/model.py
@@ -12,6 +12,7 @@ from sqlalchemy.orm import Session
 from dtbase.backend.auth import authenticate_access
 from dtbase.backend.database import models
 from dtbase.backend.database.utils import db_session
+from dtbase.backend.exc import RowMissingError
 from dtbase.backend.models import (
     MessageResponse,
     Model,
@@ -20,7 +21,6 @@ from dtbase.backend.models import (
     SensorMeasureIdentifier,
     ValueType,
 )
-from dtbase.core.exc import RowMissingError
 
 router = APIRouter(
     prefix="/model",
