+++ dtbase/backend/routers/sensor.py
@@ -12,8 +12,8 @@ from sqlalchemy.orm import Session
 from dtbase.backend.auth import authenticate_access
 from dtbase.backend.database import sensor_locations, sensors
 from dtbase.backend.database.utils import db_session
+from dtbase.backend.exc import RowMissingError
 from dtbase.backend.models import MessageResponse, SensorMeasure, SensorType, ValueType
-from dtbase.core.exc import RowMissingError
 
 router = APIRouter(
     prefix="/sensor",
